Move upload tests into app_test
